Move files to be consistent
