+++ src/components/TableView.jsx
@@ -2,7 +2,7 @@ import PropTypes from 'prop-types'
 import { useEffect, useMemo, useState } from 'react'
 import { usePagination, useSortBy, useTable } from 'react-table'
 import styled from 'styled-components'
-import ContentPageLayout from './Layout/subLayouts/ContentPageLayout/ContentPageLayout'
+import ContentPageLayout from './ContentPageLayout'
 import { getTableColumnHeaderProps } from '../library/getTableColumnHeaderProps'
 import PageSelector from './generic/Table/PageSelector/PageSelector'
 import PageSizeSelector from './generic/Table/PageSizeSelector/PageSizeSelector'
